Tutorial for the entire process
This is actually my first time to construct README.md. I gave detailed description from registering to EarthData and accessing to MERRA-2 information to subsetting and downloading them. It is kind of an introduction to what you can do afterthat with Python code to combine these .nc files into a one then converting it to a .csv.